+++ Airport.h
@@ -6,7 +6,7 @@ struct Route
 {
     unsigned int startID;
     unsigned int endID;
-    double dist;
+    long double dist;
     bool isLabeled;
 };
 
@@ -19,12 +19,12 @@ class Airport
         string iata_;
         string icao_;
         unsigned int id_;
-        double latitude_;
-        double longitude_;
+        long double latitude_;
+        long double longitude_;
     
     public:
         Airport();
-        Airport(string name, string country, string city, unsigned int id, double latitude, double longitude);
+        Airport(string name, string country, string city, unsigned int id, long double latitude, long double longitude);
         ~Airport();
 
         void setName(string name);
@@ -33,8 +33,8 @@ class Airport
         void setIATA(string iata);
         void setICAO(string icao);
         void setId(unsigned int id);
-        void setLatitude(double latitude);
-        void setLongitude(double longitude);
+        void setLatitude(long double latitude);
+        void setLongitude(long double longitude);
 
         string getName() const;
         string getCountry() const;
@@ -42,6 +42,6 @@ class Airport
         string getIATA() const;
         string getICAO() const;
         unsigned int getID() const;
-        double getLatitude() const;
-        double getLongitude() const;
+        long double getLatitude() const;
+        long double getLongitude() const;
 };
